refactor(daemon): split ServiceManager responsibilities
