+++ src/daemon/components/ServiceManager.cpp
@@ -25,7 +25,6 @@
 #include <system_error>
 #include <unordered_map>
 #include <vector>
-#include <openssl/rand.h>
 #include <yams/common/fs_utils.h>
 #include <yams/config/config_helpers.h>
 #include <yams/config/config_migration.h>
@@ -47,9 +46,6 @@
 #endif
 
 #include "../../../include/yams/daemon/components/ServiceManager.h"
-#include <boost/asio/as_tuple.hpp>
-#include <boost/asio/associated_executor.hpp>
-#include <boost/asio/async_result.hpp>
 #include <boost/asio/awaitable.hpp>
 #include <boost/asio/co_spawn.hpp>
 #include <boost/asio/detached.hpp>
@@ -59,10 +55,8 @@
 #include <boost/asio/this_coro.hpp>
 #include <boost/asio/thread_pool.hpp>
 #include <boost/asio/use_future.hpp>
-#include <tl/expected.hpp>
 #include <yams/api/content_store_builder.h>
 #include <yams/app/services/services.hpp>
-#include <yams/app/services/session_service.hpp>
 #include <yams/common/fs_utils.h>
 #include <yams/compat/thread_stop_compat.h>
 #include <yams/config/config_helpers.h>
@@ -71,23 +65,19 @@
 #include <yams/daemon/components/CheckpointManager.h>
 #include <yams/daemon/components/ConfigResolver.h>
 #include <yams/daemon/components/DaemonLifecycleFsm.h>
-#include <yams/daemon/components/DaemonMetrics.h>
 #include <yams/daemon/components/DatabaseManager.h>
 #include <yams/daemon/components/db_integrity_stamp.h>
 #include <yams/daemon/components/db_recovery.h>
 #include <yams/daemon/components/db_salvage.h>
 #include <yams/daemon/components/dispatch_utils.hpp>
 #include <yams/daemon/components/EmbeddingService.h>
-#include <yams/daemon/components/EntityGraphService.h>
 #include <yams/daemon/components/gliner_query_extractor.h>
 #include <yams/daemon/components/GraphComponent.h>
 #include <yams/daemon/components/IngestService.h>
 #include <yams/daemon/components/init_utils.hpp>
-#include <yams/daemon/components/InternalEventBus.h>
 #include <yams/daemon/components/PluginManager.h>
 #include <yams/daemon/components/ResourceGovernor.h>
 #include <yams/daemon/components/StateComponent.h>
-#include <yams/daemon/components/TopologyTuner.h>
 #include <yams/daemon/components/TuneAdvisor.h>
 #include <yams/daemon/components/VectorIndexCoordinator.h>
 #include <yams/daemon/components/VectorSystemManager.h>
@@ -95,33 +85,18 @@
 #include <yams/daemon/ipc/retrieval_session.h>
 #include <yams/daemon/metric_keys.h>
 #include <yams/daemon/shutdown_budget.h>
-#include <yams/memory_sync/content_blob_sync.h>
-#include <yams/memory_sync/memory_sync_config.h>
-#include <yams/memory_sync/memory_sync_service.h>
-#include <yams/memory_sync/records.h>
-#include <yams/metadata/metadata_sync_adapter.h>
-#include <yams/metadata/topology_sync_adapter.h>
-#include <yams/topology/topology_factory.h>
-#include <yams/vector/vector_sync_adapter.h>
-
 #include <yams/daemon/components/RepairService.h>
-#include <yams/daemon/resource/abi_content_extractor_adapter.h>
 #include <yams/daemon/resource/abi_model_provider_adapter.h>
-#include <yams/daemon/resource/abi_symbol_extractor_adapter.h>
 #include <yams/daemon/resource/external_plugin_host.h>
 #include <yams/daemon/resource/model_provider.h>
 #include <yams/daemon/resource/plugin_host.h>
 #include <yams/daemon/resource/simeon_model_provider.h>
 #include <yams/extraction/builtin_text_content_extractor.h>
-#include <yams/extraction/extraction_util.h>
 #include <yams/integrity/repair_manager.h>
 #include <yams/metadata/metadata_insert_writer.h>
 #include <yams/metadata/migration.h>
-#include <yams/plugins/symbol_extractor_v1.h>
-#include <yams/repair/embedding_repair_util.h>
-#include <yams/search/search_engine_builder.h>
+#include <yams/storage/corpus_stats.h>
 #include <yams/storage/storage_runtime_resolver.h>
-#include <yams/vector/sqlite_vec_backend.h>
 #include <yams/vector/vector_database.h>
 
 namespace {
@@ -149,10 +124,6 @@ bool isEphemeralDataDir(const std::filesystem::path& path) {
     const std::string generic = normalized.generic_string();
     return generic == "/tmp" || generic.rfind("/tmp/", 0) == 0 || generic == "/private/tmp" ||
            generic.rfind("/private/tmp/", 0) == 0;
-}
-
-std::string systemErrorMessage(int code) {
-    return std::error_code(code, std::generic_category()).message();
 }
 
 // Convenience alias for ConfigResolver timeouts
@@ -197,585 +168,6 @@ inline void setOnnxShutdownMarker(bool enabled) {
     onnxShutdownMarker().store(enabled, std::memory_order_release);
 }
 
-constexpr std::size_t kMaxP2pPrivateKeyBytes = std::size_t{64} * 1024;
-
-#ifdef _WIN32
-class WindowsTokenUser {
-public:
-    ~WindowsTokenUser() {
-        if (token_ != nullptr) {
-            CloseHandle(token_);
-        }
-    }
-
-    WindowsTokenUser(const WindowsTokenUser&) = delete;
-    WindowsTokenUser& operator=(const WindowsTokenUser&) = delete;
-    WindowsTokenUser() = default;
-
-    yams::Result<void> initialize() {
-        if (!OpenThreadToken(GetCurrentThread(), TOKEN_QUERY, TRUE, &token_)) {
-            if (GetLastError() != ERROR_NO_TOKEN ||
-                !OpenProcessToken(GetCurrentProcess(), TOKEN_QUERY, &token_)) {
-                return yams::Error{yams::ErrorCode::IOError,
-                                   "cannot open effective token for P2P identity key ACL"};
-            }
-        }
-        DWORD bytes = 0;
-        GetTokenInformation(token_, TokenUser, nullptr, 0, &bytes);
-        if (bytes == 0 || GetLastError() != ERROR_INSUFFICIENT_BUFFER) {
-            return yams::Error{yams::ErrorCode::IOError,
-                               "cannot size effective SID for P2P identity key ACL"};
-        }
-        user_.resize(bytes);
-        if (!GetTokenInformation(token_, TokenUser, user_.data(), bytes, &bytes)) {
-            return yams::Error{yams::ErrorCode::IOError,
-                               "cannot read effective SID for P2P identity key ACL"};
-        }
-        return {};
-    }
-
-    PSID sid() const { return reinterpret_cast<const TOKEN_USER*>(user_.data())->User.Sid; }
-
-private:
-    HANDLE token_{nullptr};
-    std::vector<std::byte> user_;
-};
-
-class WindowsOwnerOnlySecurity {
-public:
-    ~WindowsOwnerOnlySecurity() {
-        if (acl_ != nullptr) {
-            LocalFree(acl_);
-        }
-    }
-
-    WindowsOwnerOnlySecurity(const WindowsOwnerOnlySecurity&) = delete;
-    WindowsOwnerOnlySecurity& operator=(const WindowsOwnerOnlySecurity&) = delete;
-    WindowsOwnerOnlySecurity() = default;
-
-    yams::Result<SECURITY_ATTRIBUTES*> initialize() {
-        if (auto initialized = user_.initialize(); !initialized) {
-            return initialized.error();
-        }
-        EXPLICIT_ACCESSW access{};
-        access.grfAccessPermissions = GENERIC_ALL;
-        access.grfAccessMode = SET_ACCESS;
-        access.grfInheritance = NO_INHERITANCE;
-        BuildTrusteeWithSidW(&access.Trustee, user_.sid());
-        const auto aclResult = SetEntriesInAclW(1, &access, nullptr, &acl_);
-        if (aclResult != ERROR_SUCCESS) {
-            return yams::Error{yams::ErrorCode::IOError,
-                               "cannot build owner-only P2P identity key ACL"};
-        }
-        if (!InitializeSecurityDescriptor(&descriptor_, SECURITY_DESCRIPTOR_REVISION) ||
-            !SetSecurityDescriptorOwner(&descriptor_, user_.sid(), FALSE) ||
-            !SetSecurityDescriptorDacl(&descriptor_, TRUE, acl_, FALSE) ||
-            !SetSecurityDescriptorControl(&descriptor_, SE_DACL_PROTECTED, SE_DACL_PROTECTED)) {
-            return yams::Error{yams::ErrorCode::IOError,
-                               "cannot initialize owner-only P2P identity key ACL"};
-        }
-        attributes_.nLength = sizeof(attributes_);
-        attributes_.lpSecurityDescriptor = &descriptor_;
-        attributes_.bInheritHandle = FALSE;
-        return &attributes_;
-    }
-
-private:
-    WindowsTokenUser user_;
-    PACL acl_{nullptr};
-    SECURITY_DESCRIPTOR descriptor_{};
-    SECURITY_ATTRIBUTES attributes_{};
-};
-
-class WindowsHandle {
-public:
-    explicit WindowsHandle(HANDLE handle) : handle_(handle) {}
-    ~WindowsHandle() {
-        if (handle_ != INVALID_HANDLE_VALUE) {
-            CloseHandle(handle_);
-        }
-    }
-    WindowsHandle(const WindowsHandle&) = delete;
-    WindowsHandle& operator=(const WindowsHandle&) = delete;
-    HANDLE get() const { return handle_; }
-
-private:
-    HANDLE handle_{INVALID_HANDLE_VALUE};
-};
-
-yams::Result<std::string> readProtectedP2pFile(const std::filesystem::path& keyPath,
-                                               bool ownerOnlyRead, std::size_t maxBytes) {
-    WindowsHandle file(CreateFileW(keyPath.c_str(), GENERIC_READ | READ_CONTROL, 0, nullptr,
-                                   OPEN_EXISTING,
-                                   FILE_ATTRIBUTE_NORMAL | FILE_FLAG_OPEN_REPARSE_POINT, nullptr));
-    if (file.get() == INVALID_HANDLE_VALUE) {
-        return yams::Error{yams::ErrorCode::IOError, "cannot open P2P identity key"};
-    }
-    FILE_ATTRIBUTE_TAG_INFO attributes{};
-    if (!GetFileInformationByHandleEx(file.get(), FileAttributeTagInfo, &attributes,
-                                      sizeof(attributes)) ||
-        (attributes.FileAttributes & (FILE_ATTRIBUTE_DIRECTORY | FILE_ATTRIBUTE_REPARSE_POINT)) !=
-            0 ||
-        GetFileType(file.get()) != FILE_TYPE_DISK) {
-        return yams::Error{yams::ErrorCode::Unauthorized,
-                           "P2P identity key must be a non-reparse regular file"};
-    }
-
-    WindowsTokenUser expectedUser;
-    if (auto initialized = expectedUser.initialize(); !initialized) {
-        return initialized.error();
-    }
-    PSID owner = nullptr;
-    PACL dacl = nullptr;
-    PSECURITY_DESCRIPTOR rawDescriptor = nullptr;
-    const auto status = GetSecurityInfo(file.get(), SE_FILE_OBJECT,
-                                        OWNER_SECURITY_INFORMATION | DACL_SECURITY_INFORMATION,
-                                        &owner, nullptr, &dacl, nullptr, &rawDescriptor);
-    std::unique_ptr<void, decltype(&LocalFree)> descriptor(rawDescriptor, LocalFree);
-    if (status != ERROR_SUCCESS || owner == nullptr || dacl == nullptr ||
-        !EqualSid(owner, expectedUser.sid())) {
-        return yams::Error{yams::ErrorCode::Unauthorized,
-                           "P2P identity key must have an explicit effective-user-only ACL"};
-    }
-    SECURITY_DESCRIPTOR_CONTROL control{};
-    DWORD revision = 0;
-    if (!GetSecurityDescriptorControl(rawDescriptor, &control, &revision) ||
-        (control & SE_DACL_PROTECTED) == 0) {
-        return yams::Error{yams::ErrorCode::Unauthorized,
-                           "P2P identity key ACL must be protected from inheritance"};
-    }
-    ACL_SIZE_INFORMATION info{};
-    if (!GetAclInformation(dacl, &info, sizeof(info), AclSizeInformation)) {
-        return yams::Error{yams::ErrorCode::IOError, "cannot inspect P2P identity key ACL"};
-    }
-    bool ownerAllowed = false;
-    for (DWORD index = 0; index < info.AceCount; ++index) {
-        void* rawAce = nullptr;
-        if (!GetAce(dacl, index, &rawAce) || rawAce == nullptr) {
-            return yams::Error{yams::ErrorCode::IOError,
-                               "cannot inspect P2P identity key ACL entry"};
-        }
-        const auto* header = static_cast<const ACE_HEADER*>(rawAce);
-        if (header->AceType == ACCESS_ALLOWED_ACE_TYPE) {
-            constexpr std::size_t sidOffset = offsetof(ACCESS_ALLOWED_ACE, SidStart);
-            if (header->AceSize < sidOffset + sizeof(DWORD) ||
-                (header->AceFlags & INHERITED_ACE) != 0) {
-                return yams::Error{yams::ErrorCode::Unauthorized,
-                                   "P2P identity key ACL contains an unsafe allow entry"};
-            }
-            const auto* ace = static_cast<const ACCESS_ALLOWED_ACE*>(rawAce);
-            auto* sid = const_cast<DWORD*>(&ace->SidStart);
-            if (!IsValidSid(sid) || GetLengthSid(sid) > header->AceSize - sidOffset) {
-                return yams::Error{yams::ErrorCode::Unauthorized,
-                                   "P2P trust file ACL contains an invalid SID"};
-            }
-            const bool ownerSid = EqualSid(expectedUser.sid(), sid);
-            constexpr ACCESS_MASK readable =
-                GENERIC_ALL | GENERIC_READ | FILE_GENERIC_READ | FILE_READ_DATA | FILE_EXECUTE;
-            constexpr ACCESS_MASK writable = GENERIC_ALL | GENERIC_WRITE | FILE_GENERIC_WRITE |
-                                             FILE_WRITE_DATA | FILE_APPEND_DATA | FILE_WRITE_EA |
-                                             FILE_WRITE_ATTRIBUTES | FILE_DELETE_CHILD | DELETE |
-                                             WRITE_DAC | WRITE_OWNER;
-            if ((!ownerSid && ownerOnlyRead && ace->Mask != 0) ||
-                (!ownerSid && (ace->Mask & writable) != 0)) {
-                return yams::Error{yams::ErrorCode::Unauthorized,
-                                   "P2P trust file ACL grants unsafe external access"};
-            }
-            ownerAllowed = ownerAllowed || (ownerSid && (ace->Mask & readable) != 0);
-        } else if (header->AceType != ACCESS_DENIED_ACE_TYPE &&
-                   header->AceType != ACCESS_DENIED_OBJECT_ACE_TYPE &&
-                   header->AceType != ACCESS_DENIED_CALLBACK_ACE_TYPE &&
-                   header->AceType != ACCESS_DENIED_CALLBACK_OBJECT_ACE_TYPE) {
-            return yams::Error{yams::ErrorCode::Unauthorized,
-                               "P2P identity key ACL contains an unsupported entry type"};
-        }
-    }
-    if (!ownerAllowed) {
-        return yams::Error{yams::ErrorCode::Unauthorized,
-                           "P2P identity key ACL does not grant effective-user read access"};
-    }
-
-    LARGE_INTEGER size{};
-    if (!GetFileSizeEx(file.get(), &size) || size.QuadPart <= 0 ||
-        static_cast<unsigned long long>(size.QuadPart) > maxBytes) {
-        return yams::Error{yams::ErrorCode::InvalidArgument,
-                           "P2P identity key is empty or oversized"};
-    }
-    std::string content(static_cast<std::size_t>(size.QuadPart), '\0');
-    std::size_t offset = 0;
-    while (offset < content.size()) {
-        DWORD read = 0;
-        const auto remaining =
-            std::min<std::size_t>(content.size() - offset, std::numeric_limits<DWORD>::max());
-        if (!ReadFile(file.get(), content.data() + offset, static_cast<DWORD>(remaining), &read,
-                      nullptr) ||
-            read == 0) {
-            return yams::Error{yams::ErrorCode::IOError, "cannot read complete P2P identity key"};
-        }
-        offset += read;
-    }
-    return content;
-}
-#else
-yams::Result<std::string> readProtectedP2pFile(const std::filesystem::path& keyPath,
-                                               bool ownerOnlyRead, std::size_t maxBytes) {
-    int flags = O_RDONLY;
-#ifdef O_CLOEXEC
-    flags |= O_CLOEXEC;
-#endif
-#ifdef O_NOFOLLOW
-    flags |= O_NOFOLLOW;
-#endif
-    const int file = ::open(keyPath.c_str(), flags);
-    if (file < 0) {
-        const auto openError = errno;
-        return yams::Error{
-            openError == ELOOP ? yams::ErrorCode::Unauthorized : yams::ErrorCode::IOError,
-            openError == ELOOP ? "P2P identity key must not be a symbolic link"
-                               : "cannot open P2P identity key: " + systemErrorMessage(openError)};
-    }
-    const auto closeFile = [&] { (void)::close(file); };
-    struct stat info{};
-    if (::fstat(file, &info) != 0) {
-        const auto message = systemErrorMessage(errno);
-        closeFile();
-        return yams::Error{yams::ErrorCode::IOError, "cannot inspect P2P identity key: " + message};
-    }
-    const mode_t unsafePermissions = ownerOnlyRead ? mode_t{0077} : mode_t{0022};
-    if (!S_ISREG(info.st_mode) || info.st_uid != ::geteuid() ||
-        (info.st_mode & unsafePermissions) != 0) {
-        closeFile();
-        return yams::Error{yams::ErrorCode::Unauthorized,
-                           "P2P trust file must be an effective-user-owned regular file without "
-                           "unsafe permissions"};
-    }
-    if (info.st_size <= 0 ||
-        static_cast<std::uintmax_t>(info.st_size) > static_cast<std::uintmax_t>(maxBytes)) {
-        closeFile();
-        return yams::Error{yams::ErrorCode::InvalidArgument,
-                           "P2P identity key is empty or oversized"};
-    }
-    std::string content(static_cast<std::size_t>(info.st_size), '\0');
-    std::size_t offset = 0;
-    while (offset < content.size()) {
-        const auto read = ::read(file, content.data() + offset, content.size() - offset);
-        if (read < 0 && errno == EINTR) {
-            continue;
-        }
-        if (read <= 0) {
-            const auto message = systemErrorMessage(errno);
-            closeFile();
-            return yams::Error{yams::ErrorCode::IOError,
-                               "cannot read complete P2P identity key: " + message};
-        }
-        offset += static_cast<std::size_t>(read);
-    }
-    closeFile();
-    return content;
-}
-#endif
-
-yams::Result<std::string> readProtectedP2pPrivateKey(const std::filesystem::path& keyPath) {
-    return readProtectedP2pFile(keyPath, true, kMaxP2pPrivateKeyBytes);
-}
-
-yams::Result<std::string> readProtectedP2pTrustFile(const std::filesystem::path& path,
-                                                    std::size_t maxBytes) {
-    return readProtectedP2pFile(path, false, maxBytes);
-}
-
-yams::Result<void> writeExclusiveP2pPrivateKey(const std::filesystem::path& path,
-                                               std::string_view contents) {
-#ifdef _WIN32
-    WindowsOwnerOnlySecurity security;
-    auto securityAttributes = security.initialize();
-    if (!securityAttributes) {
-        return securityAttributes.error();
-    }
-    HANDLE file = CreateFileW(path.c_str(), GENERIC_WRITE, 0, securityAttributes.value(),
-                              CREATE_NEW, FILE_ATTRIBUTE_NORMAL, nullptr);
-    if (file == INVALID_HANDLE_VALUE) {
-        const auto code = GetLastError();
-        return yams::Error{code == ERROR_FILE_EXISTS || code == ERROR_ALREADY_EXISTS
-                               ? yams::ErrorCode::ResourceBusy
-                               : yams::ErrorCode::IOError,
-                           "cannot exclusively create temporary P2P identity key"};
-    }
-    DWORD written = 0;
-    const bool writeOk =
-        contents.size() <= std::numeric_limits<DWORD>::max() &&
-        WriteFile(file, contents.data(), static_cast<DWORD>(contents.size()), &written, nullptr) &&
-        written == contents.size() && FlushFileBuffers(file);
-    const bool closeOk = CloseHandle(file);
-    if (!writeOk || !closeOk) {
-        std::error_code ignored;
-        std::filesystem::remove(path, ignored);
-        return yams::Error{yams::ErrorCode::WriteError, "cannot write P2P identity key"};
-    }
-#else
-    int flags = O_WRONLY | O_CREAT | O_EXCL;
-#ifdef O_CLOEXEC
-    flags |= O_CLOEXEC;
-#endif
-#ifdef O_NOFOLLOW
-    flags |= O_NOFOLLOW;
-#endif
-    const int file = ::open(path.c_str(), flags, S_IRUSR | S_IWUSR);
-    if (file < 0) {
-        const int openError = errno;
-        return yams::Error{openError == EEXIST ? yams::ErrorCode::ResourceBusy
-                                               : yams::ErrorCode::IOError,
-                           "cannot exclusively create temporary P2P identity key: " +
-                               systemErrorMessage(openError)};
-    }
-    std::size_t offset = 0;
-    while (offset < contents.size()) {
-        const auto written = ::write(file, contents.data() + offset, contents.size() - offset);
-        if (written < 0 && errno == EINTR) {
-            continue;
-        }
-        if (written <= 0) {
-            const auto message = systemErrorMessage(errno);
-            (void)::close(file);
-            std::error_code ignored;
-            std::filesystem::remove(path, ignored);
-            return yams::Error{yams::ErrorCode::WriteError,
-                               "cannot write P2P identity key: " + message};
-        }
-        offset += static_cast<std::size_t>(written);
-    }
-    const bool synced = ::fsync(file) == 0;
-    const bool closed = ::close(file) == 0;
-    if (!synced || !closed) {
-        std::error_code ignored;
-        std::filesystem::remove(path, ignored);
-        return yams::Error{yams::ErrorCode::WriteError, "cannot finalize P2P identity key"};
-    }
-#endif
-    return {};
-}
-
-yams::Result<bool> installP2pPrivateKey(const std::filesystem::path& temporary,
-                                        const std::filesystem::path& keyPath) {
-#ifdef _WIN32
-    if (MoveFileExW(temporary.c_str(), keyPath.c_str(), MOVEFILE_WRITE_THROUGH)) {
-        return true;
-    }
-    const auto code = GetLastError();
-    std::error_code ignored;
-    std::filesystem::remove(temporary, ignored);
-    if (code == ERROR_FILE_EXISTS || code == ERROR_ALREADY_EXISTS) {
-        return false;
-    }
-    return yams::Error{yams::ErrorCode::IOError, "cannot install P2P identity key"};
-#else
-    if (::link(temporary.c_str(), keyPath.c_str()) == 0) {
-        std::error_code ignored;
-        std::filesystem::remove(temporary, ignored);
-        return true;
-    }
-    const int installError = errno;
-    std::error_code ignored;
-    std::filesystem::remove(temporary, ignored);
-    if (installError == EEXIST) {
-        return false;
-    }
-    return yams::Error{yams::ErrorCode::IOError,
-                       "cannot install P2P identity key: " + systemErrorMessage(installError)};
-#endif
-}
-
-yams::Result<std::string> loadOrCreateP2pPrivateKey(const std::filesystem::path& keyPath) {
-    std::error_code error;
-    if (std::filesystem::exists(keyPath, error)) {
-        return readProtectedP2pPrivateKey(keyPath);
-    }
-    if (error) {
-        return yams::Error{yams::ErrorCode::IOError,
-                           "cannot inspect P2P identity key: " + error.message()};
-    }
-    if (!yams::common::ensureDirectories(keyPath.parent_path(), error)) {
-        return yams::Error{yams::ErrorCode::IOError,
-                           "cannot create P2P identity directory: " + error.message()};
-    }
-    auto generated = yams::memory_sync::generateWriterKeyPair();
-    if (!generated) {
-        return generated.error();
-    }
-    std::random_device random;
-    for (int attempt = 0; attempt < 8; ++attempt) {
-        auto temporary = keyPath;
-        temporary += ".tmp." + std::to_string(static_cast<unsigned long long>(::getpid())) + "." +
-                     std::to_string(static_cast<unsigned long long>(random())) + "." +
-                     std::to_string(static_cast<unsigned long long>(random()));
-        auto written = writeExclusiveP2pPrivateKey(temporary, generated.value().privateKeyPem);
-        if (!written) {
-            if (written.error().code == yams::ErrorCode::ResourceBusy) {
-                continue;
-            }
-            return written.error();
-        }
-        auto installed = installP2pPrivateKey(temporary, keyPath);
-        if (!installed) {
-            return installed.error();
-        }
-        if (!installed.value()) {
-            return readProtectedP2pPrivateKey(keyPath);
-        }
-        return std::move(generated.value().privateKeyPem);
-    }
-    return yams::Error{yams::ErrorCode::ResourceBusy,
-                       "cannot allocate a unique temporary P2P identity key"};
-}
-
-yams::Result<std::string>
-loadP2pPrivateKey(const yams::daemon::DaemonConfig::MemorySyncPolicy& policy,
-                  const std::filesystem::path& dataDir) {
-    if (!policy.identityKeyPath.empty()) {
-        const std::filesystem::path configured(policy.identityKeyPath);
-        return loadOrCreateP2pPrivateKey(configured.is_absolute() ? configured
-                                                                  : dataDir / configured);
-    }
-    if (!policy.writerAuthManifestPath.empty()) {
-        const std::filesystem::path manifestPath(policy.writerAuthManifestPath);
-        auto bytes = readProtectedP2pTrustFile(manifestPath, std::size_t{1024} * 1024);
-        if (!bytes) {
-            return bytes.error();
-        }
-        try {
-            const auto manifest = nlohmann::json::parse(bytes.value());
-            if (manifest.at("schema_version").get<std::uint32_t>() != 1 ||
-                manifest.at("corpus_id").get<std::string>() != policy.corpusId ||
-                manifest.at("corpus_epoch").get<std::uint64_t>() != policy.corpusEpoch) {
-                return yams::Error{yams::ErrorCode::InvalidArgument,
-                                   "writer authentication manifest corpus or epoch mismatch"};
-            }
-            const auto& local = manifest.at("local_key");
-            if (local.at("writer_id").get<std::string>() != policy.nodeId) {
-                return yams::Error{yams::ErrorCode::InvalidArgument,
-                                   "writer authentication manifest local writer mismatch"};
-            }
-            std::filesystem::path keyPath(local.at("private_key_path").get<std::string>());
-            if (!keyPath.is_absolute()) {
-                keyPath = manifestPath.parent_path() / keyPath;
-            }
-            return readProtectedP2pPrivateKey(keyPath);
-        } catch (const std::exception&) {
-            return yams::Error{yams::ErrorCode::InvalidArgument,
-                               "writer authentication manifest is malformed"};
-        }
-    }
-    return loadOrCreateP2pPrivateKey(dataDir / "p2p" / "identity.pem");
-}
-
-yams::Result<std::string> ensureAuthenticatedDirectStore(const std::filesystem::path& dataDir,
-                                                         std::string_view corpusId,
-                                                         std::uint64_t corpusEpoch) {
-    const auto p2pDirectory = dataDir / "p2p";
-    const auto operationStore = p2pDirectory / "op-store";
-    const auto marker = p2pDirectory / "authenticated-store-v1.json";
-    std::error_code error;
-    if (std::filesystem::exists(marker, error)) {
-        if (error) {
-            return yams::Error{yams::ErrorCode::IOError,
-                               "cannot inspect direct P2P authentication marker: " +
-                                   error.message()};
-        }
-        auto encoded = readProtectedP2pPrivateKey(marker);
-        if (!encoded) {
-            return encoded.error();
-        }
-        try {
-            const auto parsed = nlohmann::json::parse(encoded.value());
-            const auto storeGeneration = parsed.at("store_generation").get<std::string>();
-            const bool validGeneration =
-                storeGeneration.size() == 64 &&
-                std::ranges::all_of(storeGeneration, [](unsigned char ch) {
-                    return (ch >= '0' && ch <= '9') || (ch >= 'a' && ch <= 'f');
-                });
-            if (parsed.at("schema_version").get<std::uint32_t>() != 1 ||
-                parsed.at("corpus_id").get<std::string>() != corpusId ||
-                parsed.at("corpus_epoch").get<std::uint64_t>() != corpusEpoch || !validGeneration) {
-                return yams::Error{
-                    yams::ErrorCode::InvalidState,
-                    "direct P2P operation store authentication marker does not match corpus "
-                    "identity"};
-            }
-            return storeGeneration;
-        } catch (const std::exception&) {
-            return yams::Error{yams::ErrorCode::DataCorruption,
-                               "direct P2P operation store authentication marker is malformed"};
-        }
-    }
-    if (error) {
-        return yams::Error{yams::ErrorCode::IOError,
-                           "cannot inspect direct P2P authentication marker: " + error.message()};
-    }
-
-    if (std::filesystem::exists(operationStore, error)) {
-        if (error) {
-            return yams::Error{yams::ErrorCode::IOError,
-                               "cannot inspect direct P2P operation store: " + error.message()};
-        }
-        const auto operationStoreStatus = std::filesystem::symlink_status(operationStore, error);
-        if (error || !std::filesystem::is_directory(operationStoreStatus)) {
-            return yams::Error{yams::ErrorCode::InvalidState,
-                               "direct P2P operation store is not a trusted directory"};
-        }
-        std::filesystem::recursive_directory_iterator entry(operationStore, error);
-        const std::filesystem::recursive_directory_iterator end;
-        for (; !error && entry != end; entry.increment(error)) {
-            const auto status = entry->symlink_status(error);
-            if (error || !std::filesystem::is_directory(status)) {
-                return yams::Error{
-                    yams::ErrorCode::InvalidState,
-                    "direct P2P operation store has unsigned or provenance-unknown history; "
-                    "preserve it for audit, advance corpus_epoch, and bootstrap a fresh store"};
-            }
-        }
-        if (error) {
-            return yams::Error{yams::ErrorCode::IOError,
-                               "cannot scan direct P2P operation store: " + error.message()};
-        }
-    } else if (error) {
-        return yams::Error{yams::ErrorCode::IOError,
-                           "cannot inspect direct P2P operation store: " + error.message()};
-    }
-
-    std::filesystem::create_directories(p2pDirectory, error);
-    if (error) {
-        return yams::Error{yams::ErrorCode::IOError,
-                           "cannot create direct P2P state directory: " + error.message()};
-    }
-    std::array<unsigned char, 32> generationBytes{};
-    if (RAND_bytes(generationBytes.data(), static_cast<int>(generationBytes.size())) != 1) {
-        return yams::Error{yams::ErrorCode::InternalError,
-                           "cannot generate direct P2P store identity"};
-    }
-    static constexpr char kHex[] = "0123456789abcdef";
-    std::string storeGeneration;
-    storeGeneration.reserve(generationBytes.size() * 2);
-    for (const auto byte : generationBytes) {
-        storeGeneration.push_back(kHex[byte >> 4]);
-        storeGeneration.push_back(kHex[byte & 0x0f]);
-    }
-    const nlohmann::json content{{"schema_version", 1},
-                                 {"corpus_id", corpusId},
-                                 {"corpus_epoch", corpusEpoch},
-                                 {"store_generation", storeGeneration}};
-    const auto temporary = marker.string() + ".tmp-" + std::to_string(::getpid());
-    if (auto written = writeExclusiveP2pPrivateKey(temporary, content.dump()); !written) {
-        return written.error();
-    }
-    std::filesystem::rename(temporary, marker, error);
-    if (error) {
-        std::filesystem::remove(temporary, error);
-        return yams::Error{yams::ErrorCode::WriteError,
-                           "cannot install direct P2P authentication marker"};
-    }
-    return storeGeneration;
-}
-
 std::uint64_t nowUnixMillis() {
     return static_cast<std::uint64_t>(std::chrono::duration_cast<std::chrono::milliseconds>(
                                           std::chrono::system_clock::now().time_since_epoch())
@@ -789,30 +181,10 @@ std::uint64_t nowUnixMillis() {
 
 namespace yams::daemon {
 
-Result<std::string>
-ServiceManager::__test_loadOrCreateP2pPrivateKey(const std::filesystem::path& keyPath) {
-    return loadOrCreateP2pPrivateKey(keyPath);
-}
-
-Result<void>
-ServiceManager::__test_writeProtectedP2pPrivateKey(const std::filesystem::path& keyPath,
-                                                   std::string_view contents) {
-    return writeExclusiveP2pPrivateKey(keyPath, contents);
-}
-
 namespace {
 constexpr auto kTopologyOverlayRebuildMinAge = std::chrono::minutes(5);
 constexpr std::size_t kTopologyOverlayDirtyThreshold = 64;
 
-/// A document whose blob cannot be replicated because its stored content is
-/// unrecoverable (corrupt object manifest, digest mismatch, malformed stored
-/// bytes) is skipped permanently: retrying the same bytes cannot succeed. Any
-/// other failure (backend unreachable, resource limits, cancellation, identity
-/// misconfiguration) propagates so the document is retried on a later cycle.
-bool isUnrecoverableBlobFailure(yams::ErrorCode code) {
-    return code == yams::ErrorCode::ManifestInvalid || code == yams::ErrorCode::HashMismatch ||
-           code == yams::ErrorCode::InvalidData;
-}
 } // namespace
 
 using yams::Error;
@@ -959,147 +331,7 @@ ServiceManager::ServiceManager(const DaemonConfig& config, StateComponent& state
         tuningLifecycleLease_.commitInitialization();
     }
 
-    {
-        auto enginePolicy = ConfigResolver::resolveTopologyEnginePolicy();
-        if (enginePolicy.engine) {
-            const auto resolved = std::string{topology::resolveFactoryKey(*enginePolicy.engine)};
-            tuningConfig_.topologyAlgorithm = resolved;
-            spdlog::info("Topology engine applied via config: {} (resolved={})",
-                         *enginePolicy.engine, resolved);
-        }
-        if (enginePolicy.routingRepresentativeCount) {
-            topologyManager_.setRoutingRepresentativeCount(
-                *enginePolicy.routingRepresentativeCount);
-            spdlog::info("Topology routing representatives applied via config: {}",
-                         topologyManager_.routingRepresentativeCount());
-        }
-        if (enginePolicy.boundarySpillEnabled) {
-            topologyManager_.setBoundarySpillPolicy(
-                *enginePolicy.boundarySpillEnabled, enginePolicy.boundarySpillLimit.value_or(1),
-                enginePolicy.boundarySpillDistanceRatio.value_or(1.05),
-                enginePolicy.boundarySpillResidualPenalty.value_or(1.0));
-            spdlog::info("Topology SOAR boundary spill applied via config: enabled={}",
-                         topologyManager_.boundarySpillEnabled());
-        }
-        topology::FeatureComposition featureComposition;
-        featureComposition.enableEntityFusion =
-            enginePolicy.featureEntityFusion.value_or(featureComposition.enableEntityFusion);
-        featureComposition.entitySignatureK =
-            enginePolicy.featureEntitySignatureK.value_or(featureComposition.entitySignatureK);
-        featureComposition.entityFusionAlpha =
-            enginePolicy.featureEntityFusionAlpha.value_or(featureComposition.entityFusionAlpha);
-        featureComposition.entityMinConfidence = enginePolicy.featureEntityMinConfidence.value_or(
-            featureComposition.entityMinConfidence);
-        featureComposition.enableMatryoshkaCoarseView =
-            enginePolicy.featureMatryoshkaCoarseView.value_or(
-                featureComposition.enableMatryoshkaCoarseView);
-        featureComposition.matryoshkaTargetDim = enginePolicy.featureMatryoshkaTargetDim.value_or(
-            featureComposition.matryoshkaTargetDim);
-        featureComposition.enableMinHashSketch =
-            enginePolicy.featureMinHashSketch.value_or(featureComposition.enableMinHashSketch);
-        featureComposition.minhashSketchDim =
-            enginePolicy.featureMinHashSketchDim.value_or(featureComposition.minhashSketchDim);
-        featureComposition.minhashAlpha =
-            enginePolicy.featureMinHashAlpha.value_or(featureComposition.minhashAlpha);
-        topologyManager_.setFeatureComposition(std::move(featureComposition));
-    }
-
-    // Audit-fix #1: throttle topology rebuild scheduling. During bulk ingest
-    // every embedding batch fires a `requestTopologyRebuild("post_ingest_drain")`
-    // and without throttling each rebuild starts immediately after the previous
-    // one finishes — turning O(N) ingest into repeated full-corpus work.
-    // Default 60s; set to 0 to disable. Env knob
-    // `YAMS_TOPOLOGY_REBUILD_MIN_INTERVAL_MS` overrides.
-    {
-        std::int64_t throttleMs = 60'000; // 60s default
-        if (const std::string raw =
-                yams::config::getenv_copy("YAMS_TOPOLOGY_REBUILD_MIN_INTERVAL_MS");
-            !raw.empty()) {
-            try {
-                throttleMs = std::max<std::int64_t>(0, std::stoll(raw));
-            } catch (const std::exception& e) {
-                spdlog::debug("Invalid YAMS_TOPOLOGY_REBUILD_MIN_INTERVAL_MS '{}': {}", raw,
-                              e.what());
-            } catch (...) {
-                spdlog::debug("Invalid YAMS_TOPOLOGY_REBUILD_MIN_INTERVAL_MS '{}': unknown "
-                              "error",
-                              raw);
-            }
-        }
-        topologyManager_.setRebuildMinIntervalMs(throttleMs);
-        spdlog::info("[ServiceManager] TopologyManager rebuild throttle = {} ms", throttleMs);
-    }
-
-    // Phase G: optional adaptive topology tuner. Disabled by default;
-    // opt-in via [topology.tuner].enabled=true in the daemon config.
-    {
-        auto tunerPolicy = ConfigResolver::resolveTopologyTunerPolicy();
-        const bool tunerEnabled = tunerPolicy.enabled.value_or(false);
-        if (tunerEnabled) {
-            TopologyTunerConfig tcfg;
-            tcfg.enabled = true;
-            if (tunerPolicy.cooldownMinutes) {
-                tcfg.cooldown = std::chrono::minutes{*tunerPolicy.cooldownMinutes};
-            }
-            if (tunerPolicy.docCountDelta) {
-                tcfg.docCountDelta = *tunerPolicy.docCountDelta;
-            }
-            if (tunerPolicy.rewardAlphaSingleton) {
-                tcfg.weights.alphaSingleton = *tunerPolicy.rewardAlphaSingleton;
-            }
-            if (tunerPolicy.rewardBetaGiantCluster) {
-                tcfg.weights.betaGiantCluster = *tunerPolicy.rewardBetaGiantCluster;
-            }
-            if (tunerPolicy.rewardGammaGiniDeviation) {
-                tcfg.weights.gammaGiniDeviation = *tunerPolicy.rewardGammaGiniDeviation;
-            }
-            if (tunerPolicy.rewardDeltaIntraEdge) {
-                tcfg.weights.deltaIntraEdge = *tunerPolicy.rewardDeltaIntraEdge;
-            }
-            // Phase H-TDA: reward mode (geometric / persistence / hybrid).
-            if (tunerPolicy.rewardMode) {
-                std::string mode = *tunerPolicy.rewardMode;
-                std::transform(mode.begin(), mode.end(), mode.begin(),
-                               [](unsigned char c) { return std::tolower(c); });
-                if (mode == "persistence") {
-                    tcfg.rewardMode = TunerRewardMode::Persistence;
-                } else if (mode == "hybrid") {
-                    tcfg.rewardMode = TunerRewardMode::Hybrid;
-                } else {
-                    tcfg.rewardMode = TunerRewardMode::Geometric;
-                }
-            }
-            if (tunerPolicy.persistenceSampleSize) {
-                tcfg.persistenceSampleSize = *tunerPolicy.persistenceSampleSize;
-            }
-            // Persist MAB state under data_dir so arm-pull history survives
-            // daemon restarts. Without this, UCB1 always picks the alphabetically
-            // first arm on each fresh daemon spawn (no pulls, infinite UCB).
-            if (!config_.dataDir.empty()) {
-                tcfg.statePath = config_.dataDir / "topology_tuner_state.json";
-            }
-            // V1: arm grid is built for a typical corpus (~5k docs); adaptive
-            // resizing as the corpus grows is a future iteration.
-            constexpr std::size_t kInitialCorpusEstimate = 5000;
-            auto tuner = std::make_shared<TopologyTuner>(tcfg);
-            tuner->setArms(defaultArmGrid(kInitialCorpusEstimate));
-            if (tcfg.statePath) {
-                if (auto r = tuner->loadState(*tcfg.statePath); !r) {
-                    spdlog::debug("[ServiceManager] topology tuner: no prior state at {} ({})",
-                                  tcfg.statePath->string(), r.error().message);
-                } else {
-                    spdlog::info("[ServiceManager] topology tuner loaded state from {}",
-                                 tcfg.statePath->string());
-                }
-            }
-            topologyManager_.setTopologyTuner(tuner);
-            spdlog::info("[ServiceManager] topology tuner enabled (cooldown={}min "
-                         "doc_delta={} arms={} state={})",
-                         std::chrono::duration_cast<std::chrono::minutes>(tcfg.cooldown).count(),
-                         tcfg.docCountDelta, tuner->arms().size(),
-                         tcfg.statePath ? tcfg.statePath->string() : std::string{"<ephemeral>"});
-        }
-    }
+    configureTopologyRuntime();
 
     metricsPublisher_.setWorkerTarget(1);
 
@@ -2588,995 +1820,6 @@ Result<std::filesystem::path> ServiceManager::initializeDataDirAndContentStore()
     return dataDir;
 }
 
-Result<void> ServiceManager::publishMemorySync(const std::string& key, const std::string& value) {
-    if (!memorySync_) {
-        return Error{ErrorCode::InvalidState, "memory sync service is not enabled"};
-    }
-    auto namespacedKey = memory_sync::userLogicalKey(key);
-    if (!namespacedKey) {
-        return namespacedKey.error();
-    }
-    std::vector<std::byte> bytes;
-    bytes.reserve(value.size());
-    for (const unsigned char byte : value) {
-        bytes.push_back(static_cast<std::byte>(byte));
-    }
-    return memorySync_->publish(namespacedKey.value(), bytes);
-}
-
-Result<void> ServiceManager::deleteMemorySync(const std::string& key) {
-    if (!memorySync_) {
-        return Error{ErrorCode::InvalidState, "memory sync service is not enabled"};
-    }
-    auto namespacedKey = memory_sync::userLogicalKey(key);
-    if (!namespacedKey) {
-        return namespacedKey.error();
-    }
-    return memorySync_->erase(namespacedKey.value(), key);
-}
-
-Result<std::string> ServiceManager::readMemorySyncCached(const std::string& key) const {
-    if (!memorySync_) {
-        return Error{ErrorCode::InvalidState, "memory sync service is not enabled"};
-    }
-    auto namespacedKey = memory_sync::userLogicalKey(key);
-    if (!namespacedKey) {
-        return namespacedKey.error();
-    }
-    auto value = memorySync_->readCached(namespacedKey.value());
-    if (!value) {
-        return value.error();
-    }
-    const auto& bytes = value.value();
-    return std::string(reinterpret_cast<const char*>(bytes.data()), bytes.size());
-}
-
-Result<ServiceManager::MemorySyncStatus> ServiceManager::getMemorySyncStatus() const {
-    if (!memorySync_) {
-        return Error{ErrorCode::InvalidState, "memory sync service is not enabled"};
-    }
-    std::uint64_t peerCount = 0;
-    if (p2pManager_) {
-        auto peers = p2pManager_->peers();
-        if (peers) {
-            peerCount = peers.value().size();
-        } else {
-            // A transient peer-registry read failure must not fail the whole status IPC; degrade
-            // to peerCount=0 so the daemon stays observable while the registry recovers.
-            spdlog::warn("[ServiceManager] p2p peer registry read failed for status: {}",
-                         peers.error().message);
-        }
-    }
-    return MemorySyncStatus{
-        memorySync_->started(),
-        memorySync_->mergedRecordCount(),
-        memorySync_->quarantinedRecordCount(),
-        memorySync_->authFailureCount(),
-        memorySync_->successfulSyncCycles(),
-        memorySync_->failedSyncCycles(),
-        memorySync_->lastSuccessfulSyncAgeMs(),
-        config_.memorySync.transport == "direct" ? "direct" : config_.memorySync.backend,
-        config_.memorySync.nodeId,
-        config_.memorySync.corpusId,
-        config_.memorySync.corpusEpoch,
-        config_.memorySync.mode,
-        config_.memorySync.transport == "direct"
-            ? (config_.memorySync.allowFirstContact ? "mutual-tls-legacy-tofu"
-                                                    : "mutual-tls-operator-pinned")
-            : (config_.memorySync.writerAuthRequired ? "authenticated-writers" : "backend-acl"),
-        peerCount};
-}
-
-Result<p2p::P2pSyncResult> ServiceManager::connectP2p(std::string_view connectionString) {
-    if (!p2pManager_) {
-        return Error{ErrorCode::InvalidState, "direct P2P transport is not enabled"};
-    }
-    return p2pManager_->connect(connectionString);
-}
-
-Result<void> ServiceManager::disconnectP2p(std::string_view nodeId) {
-    if (!p2pManager_) {
-        return Error{ErrorCode::InvalidState, "direct P2P transport is not enabled"};
-    }
-    return p2pManager_->disconnect(nodeId);
-}
-
-Result<void> ServiceManager::enrollP2pPeer(std::string_view nodeId, std::string_view spkiPin) {
-    if (!p2pManager_) {
-        return Error{ErrorCode::InvalidState, "direct P2P transport is not enabled"};
-    }
-    return p2pManager_->enrollPeer(nodeId, spkiPin);
-}
-
-Result<void> ServiceManager::forgetP2pPeer(std::string_view nodeId) {
-    if (!p2pManager_) {
-        return Error{ErrorCode::InvalidState, "direct P2P transport is not enabled"};
-    }
-    return p2pManager_->forget(nodeId);
-}
-
-Result<p2p::P2pLocalIdentity> ServiceManager::getP2pIdentity() const {
-    if (!p2pManager_) {
-        return Error{ErrorCode::InvalidState, "direct P2P transport is not enabled"};
-    }
-    return p2pManager_->localIdentity();
-}
-
-Result<std::vector<p2p::PeerRegistryRecord>> ServiceManager::listP2pPeers() const {
-    if (!p2pManager_) {
-        return Error{ErrorCode::InvalidState, "direct P2P transport is not enabled"};
-    }
-    return p2pManager_->peers();
-}
-
-Result<void> ServiceManager::stageMemorySyncDocumentDelete(std::string_view contentHash,
-                                                           bool retainContent) {
-    if (!memorySync_) {
-        return {};
-    }
-    if (!memory_sync::isSha256Digest(contentHash)) {
-        return Error{ErrorCode::InvalidArgument,
-                     "metadata deletion requires a SHA-256 content hash"};
-    }
-    auto repository = getMetadataRepo();
-    if (!repository) {
-        return Error{ErrorCode::InvalidState,
-                     "metadata repository is unavailable for replicated deletion"};
-    }
-    auto document = repository->getDocumentByHash(std::string(contentHash));
-    if (!document) {
-        return document.error();
-    }
-    if (retainContent && !document.value()) {
-        return Error{ErrorCode::InvalidArgument, "retaining content requires an existing document"};
-    }
-    const std::string documentKey =
-        std::string(memory_sync::memoryStoreName(memory_sync::MemoryStore::Document)) + "/" +
-        std::string(contentHash);
-    const std::string blobKey =
-        std::string(memory_sync::memoryStoreName(memory_sync::MemoryStore::ContentBlob)) + "/" +
-        std::string(contentHash);
-    auto probe = memory_sync::EraseReadinessProbe::MetadataAbsent;
-    if (!document.value()) {
-        auto contentStore = getContentStore();
-        if (!contentStore) {
-            return Error{ErrorCode::InvalidState,
-                         "content store is unavailable for replicated deletion"};
-        }
-        auto exists = contentStore->exists(std::string(contentHash));
-        if (!exists) {
-            return exists.error();
-        }
-        if (!exists.value()) {
-            return Error{ErrorCode::NotFound,
-                         "cannot stage replicated deletion for absent content"};
-        }
-        probe = memory_sync::EraseReadinessProbe::ContentAbsent;
-    }
-    if (retainContent) {
-        const std::array requests{
-            memory_sync::EraseStageRequest{documentKey, std::string(contentHash),
-                                           memory_sync::EraseReadinessProbe::MetadataAbsent}};
-        return memorySync_->stageErases(requests);
-    }
-
-    // Content bytes are independently addressable by hash, so standard deletion requires two
-    // durable intents. Reserve both bounded outbox slots before exposing either to the drainer.
-    const std::array requests{
-        memory_sync::EraseStageRequest{blobKey, std::string(contentHash),
-                                       memory_sync::EraseReadinessProbe::ContentAbsent},
-        memory_sync::EraseStageRequest{documentKey, std::string(contentHash), probe}};
-    return memorySync_->stageErases(requests);
-}
-
-Result<void> ServiceManager::publishMemorySyncDocumentDelete(std::string_view contentHash,
-                                                             bool retainContent) {
-    if (!memorySync_) {
-        return {};
-    }
-    std::lock_guard<std::mutex> deleteLock(memorySyncDeleteOutboxMutex_);
-    notifyMemorySyncDeleteOutboxStage("delete_publish_locked");
-    const std::string documentKey =
-        std::string(memory_sync::memoryStoreName(memory_sync::MemoryStore::Document)) + "/" +
-        std::string(contentHash);
-    const std::string blobKey =
-        std::string(memory_sync::memoryStoreName(memory_sync::MemoryStore::ContentBlob)) + "/" +
-        std::string(contentHash);
-    auto intents = memorySync_->pendingErases();
-    if (!intents) {
-        return intents.error();
-    }
-
-    bool found = false;
-    const std::vector<std::string> keys = retainContent
-                                              ? std::vector<std::string>{documentKey}
-                                              : std::vector<std::string>{blobKey, documentKey};
-    for (const auto& key : keys) {
-        const auto intent = std::ranges::find_if(
-            intents.value(), [&](const auto& candidate) { return candidate.logicalKey == key; });
-        if (intent == intents.value().end()) {
-            continue;
-        }
-        found = true;
-        const auto probe = intent->readinessProbe;
-        if (auto published = memorySync_->publishStagedErase(
-                key, [this, contentHash = std::string(contentHash),
-                      probe] { return memorySyncDeleteLocallyAbsent(contentHash, probe); });
-            !published) {
-            if (published.error().code == ErrorCode::NotFound) {
-                auto refreshed = memorySync_->pendingErases();
-                if (!refreshed) {
-                    return refreshed.error();
-                }
-                const bool stillPending =
-                    std::ranges::any_of(refreshed.value(), [&](const auto& candidate) {
-                        return candidate.logicalKey == key;
-                    });
-                if (!stillPending && memorySync_->hasCommittedTombstone(key)) {
-                    continue;
-                }
-            }
-            return published.error();
-        }
-    }
-    if (!found) {
-        if (memorySync_->hasCommittedTombstone(documentKey) &&
-            (retainContent || memorySync_->hasCommittedTombstone(blobKey))) {
-            return {};
-        }
-        return Error{ErrorCode::NotFound, "replicated deletion was not durably staged"};
-    }
-    return {};
-}
-
-Result<std::size_t> ServiceManager::applyMemorySyncContentBlobs() {
-    if (!memorySync_) {
-        return Error{ErrorCode::InvalidState, "memory sync service is not enabled"};
-    }
-    auto contentStore = getContentStore();
-    if (!contentStore) {
-        return Error{ErrorCode::InvalidState, "content store is not available"};
-    }
-
-    auto merged = memorySync_->syncOnce();
-    if (!merged) {
-        return merged.error();
-    }
-
-    memory_sync::ContentBlobSyncAdapter adapter{*contentStore, *memorySync_};
-    const std::string prefix =
-        std::string(memory_sync::memoryStoreName(memory_sync::MemoryStore::ContentBlob)) + "/";
-    std::size_t applied = 0;
-    for (const auto& [key, envelope] : merged.value()) {
-        (void)envelope;
-        if (!key.starts_with(prefix)) {
-            continue;
-        }
-        const std::string_view hash{key.data() + prefix.size(), key.size() - prefix.size()};
-        if (!memory_sync::isSha256Digest(hash)) {
-            return Error{ErrorCode::InvalidData, "invalid content-blob memory-sync key"};
-        }
-        if (envelope.isTombstone()) {
-            auto removed = adapter.applyDelete(hash);
-            if (!removed) {
-                return removed.error();
-            }
-            applied += removed.value() ? 1 : 0;
-            continue;
-        }
-        auto exists = contentStore->exists(std::string(hash));
-        if (!exists) {
-            return exists.error();
-        }
-        if (exists.value()) {
-            continue;
-        }
-        if (auto stored = adapter.applyCached(hash); !stored) {
-            return stored.error();
-        }
-        ++applied;
-    }
-    return applied;
-}
-
-void ServiceManager::notifyMemorySyncStage(std::string_view stage) noexcept {
-    std::function<void(std::string_view)> observer;
-    {
-        std::lock_guard<std::mutex> lock(memorySyncStageObserverMutex_);
-        observer = memorySyncStageObserver_;
-    }
-    if (!observer) {
-        return;
-    }
-    try {
-        observer(stage);
-    } catch (...) {
-        spdlog::warn("[ServiceManager] memory_sync stage observer threw at {}", stage);
-    }
-}
-
-void ServiceManager::notifyMemorySyncDeleteOutboxStage(std::string_view stage) noexcept {
-    std::function<void(std::string_view)> observer;
-    {
-        std::lock_guard<std::mutex> lock(memorySyncDeleteOutboxObserverMutex_);
-        observer = memorySyncDeleteOutboxObserver_;
-    }
-    if (!observer) {
-        return;
-    }
-    try {
-        observer(stage);
-    } catch (...) {
-        spdlog::warn("[ServiceManager] memory_sync delete outbox observer threw at {}", stage);
-    }
-}
-
-Result<bool>
-ServiceManager::memorySyncDeleteLocallyAbsent(std::string_view contentHash,
-                                              memory_sync::EraseReadinessProbe probe) const {
-    if (probe == memory_sync::EraseReadinessProbe::MetadataAbsent) {
-        auto repository = getMetadataRepo();
-        if (!repository) {
-            return Error{ErrorCode::InvalidState, "metadata repository is unavailable"};
-        }
-        auto document = repository->getDocumentByHash(std::string(contentHash));
-        if (!document) {
-            return document.error();
-        }
-        return !document.value().has_value();
-    }
-    if (probe == memory_sync::EraseReadinessProbe::ContentAbsent) {
-        auto contentStore = getContentStore();
-        if (!contentStore) {
-            return Error{ErrorCode::InvalidState, "content store is unavailable"};
-        }
-        auto exists = contentStore->exists(std::string(contentHash));
-        if (!exists) {
-            return exists.error();
-        }
-        return !exists.value();
-    }
-    return Error{ErrorCode::InvalidArgument,
-                 "document delete outbox requires a typed absence probe"};
-}
-
-bool ServiceManager::drainMemorySyncDocumentDeleteOutbox() noexcept {
-    try {
-        notifyMemorySyncDeleteOutboxStage("delete_drain_waiting");
-        std::lock_guard<std::mutex> deleteLock(memorySyncDeleteOutboxMutex_);
-        auto intents = memorySync_->pendingErases();
-        if (!intents) {
-            spdlog::warn("[ServiceManager] memory_sync delete outbox scan failed: {}",
-                         intents.error().message);
-            return true;
-        }
-        const std::string documentPrefix =
-            std::string(memory_sync::memoryStoreName(memory_sync::MemoryStore::Document)) + "/";
-        const std::string blobPrefix =
-            std::string(memory_sync::memoryStoreName(memory_sync::MemoryStore::ContentBlob)) + "/";
-        auto pending = std::move(intents.value());
-        std::ranges::sort(pending, [](const auto& lhs, const auto& rhs) {
-            const auto lhsCounter =
-                lhs.prepared ? lhs.preparedCounter : std::numeric_limits<std::uint64_t>::max();
-            const auto rhsCounter =
-                rhs.prepared ? rhs.preparedCounter : std::numeric_limits<std::uint64_t>::max();
-            return std::tuple{lhsCounter, lhs.logicalKey} < std::tuple{rhsCounter, rhs.logicalKey};
-        });
-        const auto validDocumentIntent = [&](const auto& intent) {
-            return intent.logicalKey == documentPrefix + intent.tombstonePayload;
-        };
-        const auto validBlobIntent = [&](const auto& intent) {
-            return intent.logicalKey == blobPrefix + intent.tombstonePayload;
-        };
-
-        bool eligible = false;
-        for (const auto& intent : pending) {
-            if (!memory_sync::isSha256Digest(intent.tombstonePayload) ||
-                (!validDocumentIntent(intent) && !validBlobIntent(intent))) {
-                continue;
-            }
-            auto absent =
-                memorySyncDeleteLocallyAbsent(intent.tombstonePayload, intent.readinessProbe);
-            if (!absent) {
-                spdlog::warn("[ServiceManager] memory_sync delete outbox probe failed: {}",
-                             absent.error().message);
-                return true;
-            }
-            if (!absent.value()) {
-                if (intent.ready) {
-                    eligible = true;
-                    if (auto cancelled = memorySync_->cancelStagedErase(intent.logicalKey);
-                        !cancelled) {
-                        spdlog::warn("[ServiceManager] stale delete outbox cancellation failed: {}",
-                                     cancelled.error().message);
-                    }
-                }
-                continue;
-            }
-            eligible = true;
-            if (auto published = memorySync_->publishStagedErase(
-                    intent.logicalKey,
-                    [this, contentHash = intent.tombstonePayload, probe = intent.readinessProbe] {
-                        return memorySyncDeleteLocallyAbsent(contentHash, probe);
-                    });
-                !published) {
-                spdlog::warn("[ServiceManager] memory_sync delete outbox retry failed: {}",
-                             published.error().message);
-            }
-        }
-        return eligible;
-    } catch (const std::exception& error) {
-        spdlog::warn("[ServiceManager] memory_sync delete outbox threw: {}", error.what());
-        return true;
-    } catch (...) {
-        spdlog::warn("[ServiceManager] memory_sync delete outbox threw (unknown)");
-        return true;
-    }
-}
-
-void ServiceManager::applyMemorySyncWinners() noexcept {
-    memorySyncApplyAttempts_.fetch_add(1, std::memory_order_acq_rel);
-    std::lock_guard<std::mutex> applyLock(memorySyncApplyMutex_);
-    if (!memorySync_ || memorySync_->stopRequested()) {
-        return;
-    }
-    // A pre-delete intent can survive a crash between local deletion and tombstone publication.
-    // Promote it only after local metadata is absent, then skip this stale callback snapshot.
-    if (drainMemorySyncDocumentDeleteOutbox()) {
-        return;
-    }
-    try {
-        // Log quarantine reasons after each cycle (the sync lock is free here, so this never
-        // blocks the IPC status path).
-        for (const auto& [key, reason] : memorySync_->quarantinedReasons()) {
-            spdlog::debug("[memory_sync] quarantined {}: {}", key, reason);
-        }
-    } catch (...) {
-        spdlog::debug("[memory_sync] quarantine reason logging skipped");
-    }
-    try {
-        if (auto result = applyMemorySyncContentBlobs(); !result) {
-            spdlog::warn("[ServiceManager] memory_sync content apply failed: {}",
-                         result.error().message);
-            return;
-        }
-    } catch (const std::exception& e) {
-        spdlog::warn("[ServiceManager] memory_sync content apply threw: {}", e.what());
-        return;
-    } catch (...) {
-        spdlog::warn("[ServiceManager] memory_sync content apply threw (unknown)");
-        return;
-    }
-    notifyMemorySyncStage("apply.after_content");
-
-    if (memorySync_->stopRequested()) {
-        return;
-    }
-    try {
-        if (auto repository = getMetadataRepo()) {
-            auto contentStore = getContentStore();
-            if (!contentStore) {
-                spdlog::warn("[ServiceManager] memory_sync metadata apply requires content store");
-                return;
-            }
-            metadata::MetadataSyncAdapter adapter{
-                *repository, *memorySync_,
-                [contentStore](std::string_view hash) {
-                    return contentStore->exists(std::string(hash));
-                },
-                [this](const metadata::DocumentInfo& document) {
-                    enqueuePostIngest(document.sha256Hash, document.mimeType);
-                }};
-            if (auto result = adapter.apply(); !result) {
-                spdlog::warn("[ServiceManager] memory_sync metadata apply failed: {}",
-                             result.error().message);
-                return;
-            }
-        }
-    } catch (const std::exception& e) {
-        spdlog::warn("[ServiceManager] memory_sync metadata apply threw: {}", e.what());
-        return;
-    } catch (...) {
-        spdlog::warn("[ServiceManager] memory_sync metadata apply threw (unknown)");
-        return;
-    }
-    notifyMemorySyncStage("apply.after_metadata");
-
-    if (memorySync_->stopRequested()) {
-        return;
-    }
-    try {
-        if (auto vectorDatabase = getVectorDatabase()) {
-            auto contentStore = getContentStore();
-            if (!contentStore) {
-                spdlog::warn("[ServiceManager] memory_sync vector apply requires content store");
-                return;
-            }
-            vector::VectorSyncAdapter::RebuildCallback rebuild;
-            if (vectorIndexCoordinator_) {
-                rebuild = [coordinator = vectorIndexCoordinator_]() {
-                    return coordinator->requestRebuildBlocking(RebuildReason::EmbeddingBatch);
-                };
-            }
-            vector::VectorSyncAdapter adapter{*vectorDatabase, *memorySync_, std::move(rebuild),
-                                              &memorySyncVectorRebuildDirty_,
-                                              [contentStore](std::string_view hash) {
-                                                  return contentStore->exists(std::string(hash));
-                                              }};
-            if (auto result = adapter.apply(); !result) {
-                spdlog::warn("[ServiceManager] memory_sync vector apply failed: {}",
-                             result.error().message);
-                return;
-            }
-        }
-    } catch (const std::exception& e) {
-        spdlog::warn("[ServiceManager] memory_sync vector apply threw: {}", e.what());
-        return;
-    } catch (...) {
-        spdlog::warn("[ServiceManager] memory_sync vector apply threw (unknown)");
-        return;
-    }
-    notifyMemorySyncStage("apply.after_vector");
-
-    if (memorySync_->stopRequested()) {
-        return;
-    }
-    try {
-        if (auto kgStore = getKgStore()) {
-            metadata::TopologySyncAdapter adapter{*kgStore, *memorySync_};
-            if (auto result = adapter.apply(); !result) {
-                spdlog::warn("[ServiceManager] memory_sync topology apply failed: {}",
-                             result.error().message);
-                return;
-            }
-        }
-    } catch (const std::exception& e) {
-        spdlog::warn("[ServiceManager] memory_sync topology apply threw: {}", e.what());
-        return;
-    } catch (...) {
-        spdlog::warn("[ServiceManager] memory_sync topology apply threw (unknown)");
-        return;
-    }
-    notifyMemorySyncStage("apply.after_topology");
-
-    if (memorySync_->stopRequested()) {
-        return;
-    }
-    const auto now = std::chrono::steady_clock::now();
-    if (now >= nextMemorySyncBackfill_) {
-        nextMemorySyncBackfill_ = now + std::chrono::seconds{5};
-        publishMemorySyncBackfill();
-    }
-}
-
-void ServiceManager::publishMemorySyncBackfill() noexcept try {
-    memorySyncBackfillAttempts_.fetch_add(1, std::memory_order_acq_rel);
-    std::lock_guard<std::mutex> backfillLock(memorySyncBackfillMutex_);
-    if (!memorySync_) {
-        return;
-    }
-
-    auto repository = getMetadataRepo();
-    auto contentStore = getContentStore();
-    auto vectorDatabase = getVectorDatabase();
-    auto kgStore = getKgStore();
-    auto& state = memorySyncBackfillState_;
-    std::size_t remainingItems = state.itemBudgetPerCycle;
-    const auto deadline = std::chrono::steady_clock::now() + state.timeBudgetPerCycle;
-    const auto shouldStop = [&] {
-        return !memorySync_ || memorySync_->stopRequested() || remainingItems == 0 ||
-               std::chrono::steady_clock::now() >= deadline;
-    };
-
-    std::size_t documentsPublished = 0;
-    std::size_t blobsPublished = 0;
-    std::size_t vectorsPublished = 0;
-    std::size_t nodesPublished = 0;
-    std::size_t edgesPublished = 0;
-    std::size_t skippedDocuments = 0;
-
-    const auto publishDocument = [&]() -> bool {
-        if (!repository || !contentStore) {
-            return false;
-        }
-        metadata::DocumentQueryOptions options;
-        options.idGreaterThan = state.documentIdCursor;
-        options.limit = 1;
-        options.orderByIdAsc = true;
-        auto documents = repository->queryDocuments(options);
-        if (!documents) {
-            throw std::runtime_error(documents.error().message);
-        }
-        if (documents.value().empty()) {
-            return false;
-        }
-
-        const auto& document = documents.value().front();
-        auto metadataValues = repository->getAllMetadata(document.id);
-        if (!metadataValues) {
-            throw std::runtime_error(metadataValues.error().message);
-        }
-        std::vector<std::pair<std::string, metadata::MetadataValue>> tags;
-        tags.reserve(metadataValues.value().size());
-        for (const auto& [key, value] : metadataValues.value()) {
-            tags.emplace_back(key, value);
-        }
-
-        memory_sync::ContentBlobSyncAdapter contentAdapter{*contentStore, *memorySync_};
-        if (!document.sha256Hash.empty()) {
-            auto published = contentAdapter.publishExisting(document.sha256Hash);
-            if (!published) {
-                // A poisoned blob (corrupt manifest, digest mismatch, malformed stored bytes)
-                // can never publish; skip it so the backfill continues with the next document
-                // instead of wedging this domain on every cycle. Transient failures propagate
-                // so the document is retried later.
-                if (isUnrecoverableBlobFailure(published.error().code)) {
-                    spdlog::warn("[ServiceManager] memory_sync backfill skipping document {} "
-                                 "(blob {} unrecoverable): {}",
-                                 document.id, document.sha256Hash, published.error().message);
-                    state.documentIdCursor = std::max(state.documentIdCursor, document.id);
-                    ++skippedDocuments;
-                    notifyMemorySyncStage("backfill.skip_document");
-                    return true;
-                }
-                throw std::runtime_error(published.error().message);
-            }
-            blobsPublished += published.value() ? 1U : 0U;
-        }
-        metadata::MetadataSyncAdapter metadataAdapter{*repository, *memorySync_};
-        auto published = metadataAdapter.publish(document, tags);
-        if (!published) {
-            if (isUnrecoverableBlobFailure(published.error().code)) {
-                spdlog::warn("[ServiceManager] memory_sync backfill skipping document {} "
-                             "(record unrecoverable): {}",
-                             document.id, published.error().message);
-                state.documentIdCursor = std::max(state.documentIdCursor, document.id);
-                ++skippedDocuments;
-                notifyMemorySyncStage("backfill.skip_document");
-                return true;
-            }
-            throw std::runtime_error(published.error().message);
-        }
-        state.documentIdCursor = std::max(state.documentIdCursor, document.id);
-        ++documentsPublished;
-        notifyMemorySyncStage("backfill.after_document");
-        return true;
-    };
-
-    const auto publishVector = [&]() -> bool {
-        if (!vectorDatabase) {
-            return false;
-        }
-        auto records = vectorDatabase->getVectorsPage(state.vectorDocumentHashCursor,
-                                                      state.vectorChunkIdCursor, 1);
-        if (!records) {
-            throw std::runtime_error(records.error().message);
-        }
-        if (records.value().empty()) {
-            return false;
-        }
-        const auto& record = records.value().front();
-        vector::VectorSyncAdapter adapter{*vectorDatabase, *memorySync_};
-        auto published = adapter.publish(record);
-        if (!published) {
-            throw std::runtime_error(published.error().message);
-        }
-        state.vectorDocumentHashCursor = record.document_hash;
-        state.vectorChunkIdCursor = record.chunk_id;
-        ++vectorsPublished;
-        notifyMemorySyncStage("backfill.after_vector");
-        return true;
-    };
-
-    const auto publishTopology = [&]() -> bool {
-        if (!kgStore) {
-            return false;
-        }
-        metadata::TopologySyncAdapter adapter{*kgStore, *memorySync_};
-        if (!state.topologySnapshotInitialized) {
-            auto nodeTypes = kgStore->getNodeTypeCounts();
-            if (!nodeTypes) {
-                throw std::runtime_error(nodeTypes.error().message);
-            }
-            state.topologyNodeTypes.reserve(nodeTypes.value().size());
-            for (const auto& [nodeType, count] : nodeTypes.value()) {
-                (void)count;
-                state.topologyNodeTypes.push_back(nodeType);
-            }
-            std::sort(state.topologyNodeTypes.begin(), state.topologyNodeTypes.end());
-            state.topologySnapshotInitialized = true;
-        }
-
-        while (!shouldStop()) {
-            if (state.topologyNodeActive) {
-                auto edges = kgStore->getEdgesFrom(state.topologyNodeId, std::nullopt, 1,
-                                                   state.topologyEdgeOffset);
-                if (!edges) {
-                    throw std::runtime_error(edges.error().message);
-                }
-                if (!edges.value().empty()) {
-                    const auto& edge = edges.value().front();
-                    auto target = kgStore->getNodeById(edge.dstNodeId);
-                    if (!target) {
-                        throw std::runtime_error(target.error().message);
-                    }
-                    ++state.topologyEdgeOffset;
-                    if (!target.value()) {
-                        continue;
-                    }
-                    auto published =
-                        adapter.publishEdge(state.topologyNodeKey, edge, target.value()->nodeKey);
-                    if (!published) {
-                        --state.topologyEdgeOffset;
-                        throw std::runtime_error(published.error().message);
-                    }
-                    ++edgesPublished;
-                    notifyMemorySyncStage("backfill.after_edge");
-                    return true;
-                }
-                state.topologyNodeActive = false;
-                state.topologyNodeId = 0;
-                state.topologyNodeKey.clear();
-                state.topologyEdgeOffset = 0;
-                ++state.topologyNodeOffsets[state.topologyNodeTypes[state.topologyTypeIndex]];
-            }
-
-            if (state.topologyTypeIndex >= state.topologyNodeTypes.size()) {
-                // Refresh the bounded type snapshot on the next cycle so node types created after
-                // startup become visible. Per-type offsets prevent replaying nodes already swept.
-                state.topologySnapshotInitialized = false;
-                state.topologyNodeTypes.clear();
-                state.topologyTypeIndex = 0;
-                return false;
-            }
-            const auto& nodeType = state.topologyNodeTypes[state.topologyTypeIndex];
-            auto nodes = kgStore->findNodesByType(nodeType, 1, state.topologyNodeOffsets[nodeType]);
-            if (!nodes) {
-                throw std::runtime_error(nodes.error().message);
-            }
-            if (nodes.value().empty()) {
-                ++state.topologyTypeIndex;
-                continue;
-            }
-
-            const auto& node = nodes.value().front();
-            auto published = adapter.publishNode(node);
-            if (!published) {
-                throw std::runtime_error(published.error().message);
-            }
-            state.topologyNodeId = node.id;
-            state.topologyNodeKey = node.nodeKey;
-            state.topologyNodeActive = true;
-            ++nodesPublished;
-            notifyMemorySyncStage("backfill.after_node");
-            return true;
-        }
-        return false;
-    };
-
-    const auto advanceDomain = [&] {
-        using Domain = MemorySyncBackfillState::Domain;
-        switch (state.nextDomain) {
-            case Domain::Documents:
-                state.nextDomain = Domain::Vectors;
-                break;
-            case Domain::Vectors:
-                state.nextDomain = Domain::Topology;
-                break;
-            case Domain::Topology:
-                state.nextDomain = Domain::Documents;
-                break;
-        }
-    };
-
-    std::size_t consecutiveEmptyDomains = 0;
-    while (!shouldStop() && consecutiveEmptyDomains < 3) {
-        const auto domain = state.nextDomain;
-        advanceDomain();
-        bool published = false;
-        try {
-            using Domain = MemorySyncBackfillState::Domain;
-            switch (domain) {
-                case Domain::Documents:
-                    published = publishDocument();
-                    break;
-                case Domain::Vectors:
-                    published = publishVector();
-                    break;
-                case Domain::Topology:
-                    published = publishTopology();
-                    break;
-            }
-        } catch (const std::exception& error) {
-            spdlog::warn("[ServiceManager] memory_sync backfill domain failed: {}", error.what());
-        } catch (...) {
-            spdlog::warn("[ServiceManager] memory_sync backfill domain failed (unknown)");
-        }
-
-        if (published) {
-            --remainingItems;
-            consecutiveEmptyDomains = 0;
-        } else {
-            ++consecutiveEmptyDomains;
-        }
-    }
-
-    spdlog::debug(
-        "[ServiceManager] memory_sync backfill scanned documents={} blobs={} vectors={} nodes={} "
-        "edges={} skipped={}",
-        documentsPublished, blobsPublished, vectorsPublished, nodesPublished, edgesPublished,
-        skippedDocuments);
-} catch (const std::exception& error) {
-    spdlog::warn("[ServiceManager] memory_sync backfill setup failed: {}", error.what());
-} catch (...) {
-    spdlog::warn("[ServiceManager] memory_sync backfill setup failed (unknown)");
-}
-
-Result<void> ServiceManager::configureMemorySyncApply() {
-    if (!memorySync_) {
-        return {};
-    }
-    memorySync_->setAfterSyncCallback([this] { applyMemorySyncWinners(); });
-    // Recover or revalidate durable local delete intents before the worker and direct-P2P
-    // listener can expose a checkpoint whose exact tombstone index is still incomplete.
-    (void)drainMemorySyncDocumentDeleteOutbox();
-    if (auto started = memorySync_->start(); !started) {
-        return Error{started.error().code,
-                     "memory_sync service failed to start: " + started.error().message};
-    }
-    spdlog::info("[ServiceManager] memory_sync apply path enabled and worker started");
-    return {};
-}
-
-Result<void> ServiceManager::initializeMemorySync(const std::filesystem::path& dataDir) {
-    const auto& policy = config_.memorySync;
-    if (!policy.enabled) {
-        return Result<void>();
-    }
-    if (policy.transport == "direct" &&
-        (!policy.writerAuthRequired || policy.writerAuthManifestPath.empty())) {
-        return Error{ErrorCode::InvalidArgument,
-                     "direct memory_sync requires writer_auth_required=true and a usable "
-                     "writer_auth_manifest for authenticated cold bootstrap"};
-    }
-
-    try {
-        yams::memory_sync::MemorySyncDaemonConfig cfg;
-        cfg.enabled = true;
-        cfg.nodeId = policy.nodeId;
-        cfg.corpusId = policy.corpusId;
-        cfg.corpusEpoch = policy.corpusEpoch;
-        cfg.backend = policy.transport == "direct" ? "filesystem" : policy.backend;
-        cfg.syncIntervalMs = policy.syncIntervalMs;
-        cfg.limits = policy.limits;
-        cfg.mode = policy.mode == "persistent-migration" ? "persistent" : policy.mode;
-        cfg.sessionId = policy.sessionId;
-        cfg.temporarySessionTtlMs = policy.temporarySessionTtlMs;
-        cfg.allowLegacyUnbound = policy.mode == "persistent-migration";
-        cfg.writerAuthRequired = policy.writerAuthRequired;
-        cfg.writerAuthManifestPath = policy.writerAuthManifestPath;
-        cfg.path =
-            policy.transport == "direct" ? (dataDir / "p2p" / "op-store").string() : policy.path;
-        if (cfg.backend == "filesystem" && !cfg.path.empty() &&
-            !std::filesystem::path(cfg.path).is_absolute()) {
-            cfg.path = (dataDir / cfg.path).string();
-        }
-
-        if (policy.transport == "direct") {
-            auto trustedStore =
-                ensureAuthenticatedDirectStore(dataDir, cfg.corpusId, cfg.corpusEpoch);
-            if (!trustedStore) {
-                return trustedStore.error();
-            }
-            cfg.controlScope = std::move(trustedStore.value());
-        }
-
-        std::optional<yams::storage::BackendConfig> resolvedS3Config;
-        if (policy.transport == "shared-store" && cfg.backend == "s3") {
-            auto storageDecision = yams::storage::resolveStorageBootstrapDecision(
-                config_.configFilePath, dataDir, std::string{"s3"});
-            if (!storageDecision) {
-                return Error{storageDecision.error().code,
-                             "memory_sync S3 credential resolution failed: " +
-                                 storageDecision.error().message};
-            }
-            if (!storageDecision.value().backendConfig) {
-                return Error{ErrorCode::InvalidState,
-                             "memory_sync S3 resolution returned no backend config"};
-            }
-            resolvedS3Config = *storageDecision.value().backendConfig;
-        }
-
-        auto svc = yams::memory_sync::createMemorySyncService(
-            cfg, std::move(resolvedS3Config),
-            [] { return ResourceGovernor::instance().canAdmitWork(); },
-            policy.transport == "direct"
-                ? yams::memory_sync::WriterPrivateKeyReader{readProtectedP2pPrivateKey}
-                : yams::memory_sync::WriterPrivateKeyReader{},
-            policy.transport == "direct"
-                ? yams::memory_sync::WriterTrustFileReader{readProtectedP2pTrustFile}
-                : yams::memory_sync::WriterTrustFileReader{});
-        if (!svc) {
-            return Error{svc.error().code,
-                         "memory_sync service creation failed: " + svc.error().message};
-        }
-
-        memorySync_ = std::move(svc.value());
-        const std::string_view trust =
-            policy.transport == "direct"
-                ? (policy.allowFirstContact ? "mutual-tls-legacy-tofu"
-                                            : "mutual-tls-operator-pinned")
-                : (cfg.writerAuthRequired ? "authenticated-writers" : "backend-ACL-trusted");
-        spdlog::info("[ServiceManager] memory_sync configured: transport={} backend={} mode={} "
-                     "corpus={} epoch={} node={} trust={} sync_interval_ms={} path={}",
-                     policy.transport, cfg.backend, cfg.mode, cfg.corpusId, cfg.corpusEpoch,
-                     cfg.nodeId, trust, cfg.syncIntervalMs, cfg.path);
-        return Result<void>();
-    } catch (const std::exception& e) {
-        return Error{ErrorCode::InternalError,
-                     std::string("memory_sync initialization threw: ") + e.what()};
-    } catch (...) {
-        return Error{ErrorCode::InternalError,
-                     "memory_sync initialization threw an unknown exception"};
-    }
-}
-
-Result<void> ServiceManager::initializeDirectP2p(const std::filesystem::path& dataDir) {
-    const auto& policy = config_.memorySync;
-    if (!policy.enabled || policy.transport != "direct") {
-        return Result<void>();
-    }
-    if (!memorySync_) {
-        return Error{ErrorCode::InvalidState, "direct P2P requires memory sync service"};
-    }
-    auto listen = p2p::parseP2pConnectionString(policy.listen);
-    if (!listen) {
-        return Error{listen.error().code, "invalid memory_sync.listen: " + listen.error().message};
-    }
-    auto privateKey = loadP2pPrivateKey(policy, dataDir);
-    if (!privateKey) {
-        return privateKey.error();
-    }
-    auto synchronized = memorySync_->syncFully();
-    if (!synchronized) {
-        return Error{synchronized.error().code,
-                     "direct P2P local op-store recovery failed: " + synchronized.error().message};
-    }
-    if (memorySync_->legacyUnauthenticatedHistoryObserved()) {
-        return Error{ErrorCode::InvalidState,
-                     "direct P2P operation store contains unsigned legacy history; preserve it "
-                     "for audit, advance corpus_epoch, and bootstrap a fresh store"};
-    }
-    if (policy.allowFirstContact) {
-        spdlog::warn("[ServiceManager] memory_sync.allow_first_contact=true permits unsolicited "
-                     "P2P peer enrollment; use operator-pinned enrollment instead");
-    }
-    auto manager = p2p::P2pManager::create(
-        p2p::P2pManagerOptions{.nodeId = policy.nodeId,
-                               .corpusId = policy.corpusId,
-                               .corpusEpoch = policy.corpusEpoch,
-                               .privateKeyPem = std::move(privateKey.value()),
-                               .databasePath = dataDir / "yams.db",
-                               .listenHost = listen.value().host,
-                               .listenPort = listen.value().port,
-                               .allowFirstContact = policy.allowFirstContact,
-                               .maxPeers = policy.maxPeers,
-                               .reconnectInterval =
-                                   std::chrono::milliseconds(policy.syncIntervalMs),
-                               .timeout = std::chrono::seconds(10)},
-        *memorySync_);
-    if (!manager) {
-        return manager.error();
-    }
-    if (auto started = manager.value()->start(); !started) {
-        return started.error();
-    }
-    spdlog::info("[ServiceManager] direct P2P listening on {}:{} with identity {}",
-                 listen.value().host, manager.value()->boundPort(), policy.nodeId);
-    p2pManager_ = std::move(manager.value());
-    return Result<void>();
-}
-
 boost::asio::awaitable<bool>
 ServiceManager::initializeMetadataDatabaseAt(const std::filesystem::path& dbPath,
                                              yams::compat::stop_token token) {
@@ -4829,203 +3072,6 @@ void ServiceManager::startDeferredMetadataWarmup() {
             spdlog::warn("[ServiceManager] Deferred metadata warmup failed");
         }
     });
-}
-
-bool ServiceManager::shouldStartSessionWatcher(std::string_view disableValue) {
-    const std::string value(disableValue);
-    return !ConfigResolver::envTruthy(value.c_str());
-}
-
-std::chrono::milliseconds ServiceManager::sessionWatcherDelay(bool watchEnabled,
-                                                              std::uint32_t intervalMs) {
-    constexpr auto kIdleDelay = std::chrono::milliseconds(2000);
-    constexpr std::uint32_t kMinimumIntervalMs = 100;
-    if (!watchEnabled) {
-        return kIdleDelay;
-    }
-    return std::chrono::milliseconds(std::max(kMinimumIntervalMs, intervalMs));
-}
-
-app::services::AddDirectoryRequest
-ServiceManager::makeSessionWatchRequest(std::string_view session,
-                                        const std::filesystem::path& directory,
-                                        std::vector<std::string> changed) {
-    app::services::AddDirectoryRequest request;
-    request.directoryPath = directory.string();
-    request.includePatterns = std::move(changed);
-    request.recursive = true;
-    request.sessionId = std::string(session);
-    request.noGitignore = false;
-    return request;
-}
-
-bool ServiceManager::scanSessionWatchDirectory(app::services::IIndexingService& indexingService,
-                                               app::services::IDocumentService* documentService,
-                                               std::string_view session,
-                                               const std::filesystem::path& directory) {
-    std::error_code error;
-    if (directory.empty() || !std::filesystem::is_directory(directory, error)) {
-        return false;
-    }
-
-    auto& previousFiles = sessionWatch_.dirFiles[directory.string()];
-    std::unordered_map<std::string, std::pair<std::uint64_t, std::uint64_t>> currentFiles;
-    std::vector<std::string> changed;
-    auto it = std::filesystem::recursive_directory_iterator(directory, error);
-    const auto end = std::filesystem::recursive_directory_iterator();
-    for (; !error && it != end; it.increment(error)) {
-        std::error_code metadataError;
-        if (!it->is_regular_file(metadataError)) {
-            if (metadataError) {
-                error = metadataError;
-            }
-            continue;
-        }
-
-        const auto filePath = it->path().string();
-        const auto fileSize = static_cast<std::uint64_t>(it->file_size(metadataError));
-        if (metadataError) {
-            error = metadataError;
-            break;
-        }
-        const auto modifiedTime = it->last_write_time(metadataError);
-        if (metadataError) {
-            error = metadataError;
-            break;
-        }
-        const auto modifiedAt = static_cast<std::uint64_t>(modifiedTime.time_since_epoch().count());
-        currentFiles[filePath] = {modifiedAt, fileSize};
-        const auto previous = previousFiles.find(filePath);
-        if (previous == previousFiles.end() || previous->second != currentFiles[filePath]) {
-            std::error_code relativeError;
-            const auto relativePath =
-                std::filesystem::relative(it->path(), directory, relativeError);
-            auto relative =
-                relativeError ? it->path().filename().string() : relativePath.generic_string();
-            if (!relative.empty()) {
-                changed.emplace_back(std::move(relative));
-            }
-        }
-    }
-    if (error) {
-        spdlog::warn("[ServiceManager] session watcher scan failed for '{}': {}",
-                     directory.string(), error.message());
-        return false;
-    }
-
-    std::vector<std::string> removed;
-    removed.reserve(previousFiles.size());
-    for (const auto& [filePath, fingerprint] : previousFiles) {
-        (void)fingerprint;
-        if (!currentFiles.contains(filePath)) {
-            removed.push_back(filePath);
-        }
-    }
-
-    if (!changed.empty()) {
-        auto request = makeSessionWatchRequest(session, directory, std::move(changed));
-        auto indexed = indexingService.addDirectory(request);
-        if (!indexed || indexed.value().filesFailed != 0) {
-            const auto detail =
-                indexed ? std::to_string(indexed.value().filesFailed) + " changed file(s) failed"
-                        : indexed.error().message;
-            spdlog::warn("[ServiceManager] session watcher indexing failed for '{}': {}",
-                         directory.string(), detail);
-            return false;
-        }
-    }
-
-    if (!removed.empty() && documentService == nullptr) {
-        spdlog::warn("[ServiceManager] session watcher cannot remove {} stale path(s) for '{}': "
-                     "document service unavailable",
-                     removed.size(), directory.string());
-        return false;
-    }
-    for (const auto& filePath : removed) {
-        app::services::DeleteByNameRequest request;
-        request.name = filePath;
-        request.force = false;
-        auto deleted = documentService->deleteByName(request);
-        if (!deleted) {
-            if (deleted.error().code == ErrorCode::NotFound) {
-                continue;
-            }
-            spdlog::warn("[ServiceManager] session watcher removal failed for '{}': {}", filePath,
-                         deleted.error().message);
-            return false;
-        }
-        if (!deleted.value().errors.empty()) {
-            spdlog::warn("[ServiceManager] session watcher removal failed for '{}': {}", filePath,
-                         deleted.value().errors.front().error.value_or("unknown error"));
-            return false;
-        }
-    }
-
-    previousFiles.swap(currentFiles);
-    return true;
-}
-
-std::chrono::milliseconds ServiceManager::runSessionWatcherIteration() {
-    yams::app::services::AppContext appCtx = getAppContext();
-    auto sessionService = yams::app::services::makeSessionService(&appCtx);
-    const auto current = sessionService->current();
-    if (!current) {
-        return sessionWatcherDelay(false, 0);
-    }
-
-    const bool watchEnabled = sessionService->watchEnabled(current);
-    const auto delay = sessionWatcherDelay(watchEnabled, sessionService->watchIntervalMs(current));
-    if (!watchEnabled) {
-        return delay;
-    }
-
-    auto indexingService = yams::app::services::makeIndexingService(appCtx);
-    if (!indexingService) {
-        return delay;
-    }
-    auto documentService = yams::app::services::makeDocumentService(appCtx);
-
-    for (const auto& pattern : sessionService->getPinnedPatterns(current)) {
-        scanSessionWatchDirectory(*indexingService, documentService.get(), *current, pattern);
-    }
-    return delay;
-}
-
-boost::asio::awaitable<void>
-ServiceManager::co_runSessionWatcher(const yams::compat::stop_token& token) {
-    auto executor = co_await boost::asio::this_coro::executor;
-    boost::asio::steady_timer timer(executor);
-
-    while (!token.stop_requested()) {
-        auto waitDuration = sessionWatcherDelay(false, 0);
-        try {
-            waitDuration = runSessionWatcherIteration();
-        } catch (const std::exception& e) {
-            spdlog::debug("[ServiceManager] session watcher iteration failed: {}", e.what());
-        } catch (...) {
-            spdlog::debug(
-                "[ServiceManager] session watcher iteration failed with unknown exception");
-        }
-
-        try {
-            if (retrievalSessions_) {
-                retrievalSessions_->cleanupExpired(std::chrono::seconds(60));
-            }
-        } catch (const std::exception& e) {
-            spdlog::debug("[ServiceManager] retrieval-session cleanup failed: {}", e.what());
-        } catch (...) {
-            spdlog::debug(
-                "[ServiceManager] retrieval-session cleanup failed with unknown exception");
-        }
-
-        boost::system::error_code ec;
-        timer.expires_after(waitDuration);
-        co_await timer.async_wait(boost::asio::redirect_error(boost::asio::use_awaitable, ec));
-        if (token.stop_requested() || ec == boost::asio::error::operation_aborted)
-            break;
-    }
-
-    co_return;
 }
 
 void ServiceManager::setDatabasePhase(std::string_view phase) {
