+++ src/daemon/meson.build
@@ -112,6 +112,10 @@ yams_daemon_lib = static_library('yams_daemon',
     'components/EmbeddingLifecycleManager.cpp',
     'components/WorkCoordinator.cpp',
     'components/ServiceManager.cpp',
+    'components/service_manager/memory_sync.cpp',
+    'components/service_manager/p2p_integration.cpp',
+    'components/service_manager/session_watcher.cpp',
+    'components/service_manager/topology_policy.cpp',
     'components/SearchEngineManager.cpp',
     'components/SearchComponent.cpp',
     'components/GraphComponent.cpp',
